% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{air_quality_in_cities}
\alias{air_quality_in_cities}
\title{air_quality_in_cities dataset}
\format{
An object of class \code{data.frame} with 1500 rows and 9 columns.
}
\usage{
air_quality_in_cities
}
\description{
Imported from \code{datasets/air_quality_in_cities.csv}.
}
\keyword{datasets}
