% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{arm_span}
\alias{arm_span}
\title{arm_span dataset}
\format{
An object of class \code{data.frame} with 90 rows and 4 columns.
}
\usage{
arm_span
}
\description{
Imported from \code{datasets/arm_span.csv}.
}
\keyword{datasets}
