% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{atu_clean}
\alias{atu_clean}
\title{atu_clean dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 8 columns.
}
\usage{
atu_clean
}
\description{
Imported from \code{datasets/atu_clean.csv}.
}
\keyword{datasets}
