% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{atu_dirty}
\alias{atu_dirty}
\title{atu_dirty dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 8 columns.
}
\usage{
atu_dirty
}
\description{
Imported from \code{datasets/atu_dirty.csv}.
}
\keyword{datasets}
