% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{atus}
\alias{atus}
\title{atus dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 43 columns.
}
\usage{
atus
}
\description{
Imported from \code{datasets/atus.csv}.
}
\keyword{datasets}
