% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{cdc}
\alias{cdc}
\title{cdc dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 33 columns.
}
\usage{
cdc
}
\description{
Imported from \code{datasets/cdc.csv}.
}
\keyword{datasets}
