% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{census}
\alias{census}
\title{census dataset}
\format{
An object of class \code{data.frame} with 1000 rows and 60 columns.
}
\usage{
census
}
\description{
Imported from \code{datasets/census.csv}.
}
\keyword{datasets}
