% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{cereal}
\alias{cereal}
\title{cereal dataset}
\format{
An object of class \code{data.frame} with 40 rows and 12 columns.
}
\usage{
cereal
}
\description{
Imported from \code{datasets/cereal.csv}.
}
\keyword{datasets}
