% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{cereal_test}
\alias{cereal_test}
\title{cereal_test dataset}
\format{
An object of class \code{data.frame} with 20 rows and 12 columns.
}
\usage{
cereal_test
}
\description{
Imported from \code{datasets/cereal_test.csv}.
}
\keyword{datasets}
