% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{college_student_placement}
\alias{college_student_placement}
\title{college_student_placement dataset}
\format{
An object of class \code{data.frame} with 1000 rows and 10 columns.
}
\usage{
college_student_placement
}
\description{
Imported from \code{datasets/college_student_placement.csv}.
}
\keyword{datasets}
