% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{diabetes dataset}
\format{
An object of class \code{data.frame} with 442 rows and 11 columns.
}
\usage{
diabetes
}
\description{
Imported from \code{datasets/diabetes.csv}.
}
\keyword{datasets}
