% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{dwp_2010}
\alias{dwp_2010}
\title{dwp_2010 dataset}
\format{
An object of class \code{data.frame} with 1705 rows and 19 columns.
}
\usage{
dwp_2010
}
\description{
Imported from \code{datasets/dwp_2010.csv}.
}
\keyword{datasets}
