% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{dwp_student}
\alias{dwp_student}
\title{dwp_student dataset}
\format{
An object of class \code{data.frame} with 1193 rows and 19 columns.
}
\usage{
dwp_student
}
\description{
Imported from \code{datasets/dwp_student.csv}.
}
\keyword{datasets}
