% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{dwp_teacher}
\alias{dwp_teacher}
\title{dwp_teacher dataset}
\format{
An object of class \code{data.frame} with 512 rows and 19 columns.
}
\usage{
dwp_teacher
}
\description{
Imported from \code{datasets/dwp_teacher.csv}.
}
\keyword{datasets}
