% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{enterprise_genai_adoption_impact}
\alias{enterprise_genai_adoption_impact}
\title{enterprise_genai_adoption_impact dataset}
\format{
An object of class \code{data.frame} with 1000 rows and 10 columns.
}
\usage{
enterprise_genai_adoption_impact
}
\description{
Imported from \code{datasets/enterprise_genai_adoption_impact.csv}.
}
\keyword{datasets}
