% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{faithful}
\alias{faithful}
\title{faithful dataset}
\format{
An object of class \code{data.frame} with 272 rows and 2 columns.
}
\usage{
faithful
}
\description{
Imported from \code{datasets/faithful.csv}.
}
\keyword{datasets}
