% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{futbol}
\alias{futbol}
\title{futbol dataset}
\format{
An object of class \code{data.frame} with 55 rows and 6 columns.
}
\usage{
futbol
}
\description{
Imported from \code{datasets/futbol.csv}.
}
\keyword{datasets}
