% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{iris}
\alias{iris}
\title{iris dataset}
\format{
An object of class \code{data.frame} with 150 rows and 5 columns.
}
\usage{
iris
}
\description{
Imported from \code{datasets/iris.csv}.
}
\keyword{datasets}
