% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{mnist}
\alias{mnist}
\title{mnist dataset}
\format{
An object of class \code{data.frame} with 5000 rows and 785 columns.
}
\usage{
mnist
}
\description{
Imported from \code{datasets/mnist.csv}.
}
\keyword{datasets}
