% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{movie}
\alias{movie}
\title{movie dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 19 columns.
}
\usage{
movie
}
\description{
Imported from \code{datasets/movie.csv}.
}
\keyword{datasets}
