% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{mtcars dataset}
\format{
An object of class \code{data.frame} with 32 rows and 12 columns.
}
\usage{
mtcars
}
\description{
Imported from \code{datasets/mtcars.csv}.
}
\keyword{datasets}
