% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{mustang}
\alias{mustang}
\title{mustang dataset}
\format{
An object of class \code{data.frame} with 25 rows and 3 columns.
}
\usage{
mustang
}
\description{
Imported from \code{datasets/mustang.csv}.
}
\keyword{datasets}
