% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{neighborhoods}
\alias{neighborhoods}
\title{neighborhoods dataset}
\format{
An object of class \code{data.frame} with 100 rows and 5 columns.
}
\usage{
neighborhoods
}
\description{
Imported from \code{datasets/neighborhoods.csv}.
}
\keyword{datasets}
