% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{nutrition}
\alias{nutrition}
\title{nutrition dataset}
\format{
An object of class \code{data.frame} with 1241 rows and 12 columns.
}
\usage{
nutrition
}
\description{
Imported from \code{datasets/nutrition.csv}.
}
\keyword{datasets}
