% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{penguins}
\alias{penguins}
\title{penguins dataset}
\format{
An object of class \code{data.frame} with 344 rows and 8 columns.
}
\usage{
penguins
}
\description{
Imported from \code{datasets/penguins.csv}.
}
\keyword{datasets}
