% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{plant_disease}
\alias{plant_disease}
\title{plant_disease dataset}
\format{
An object of class \code{data.frame} with 999 rows and 5 columns.
}
\usage{
plant_disease
}
\description{
Imported from \code{datasets/plant_disease.csv}.
}
\keyword{datasets}
