% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonly-package.R
\docType{package}
\name{pythonly-package}
\alias{pythonly}
\alias{pythonly-package}
\title{pythonly: Built-in datasets for Pythonly}
\description{
The \code{pythonly} package provides datasets (packaged from CSV files) for use in
Pythonly curricula and examples.
}
\author{
\strong{Maintainer}: Pythonly Contributors \email{info@pythonly.org}

Authors:
\itemize{
  \item Pythonly Contributors \email{info@pythonly.org}
}

}
\keyword{internal}
