% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{quakes}
\alias{quakes}
\title{quakes dataset}
\format{
An object of class \code{data.frame} with 1000 rows and 5 columns.
}
\usage{
quakes
}
\description{
Imported from \code{datasets/quakes.csv}.
}
\keyword{datasets}
