% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{quickdraw}
\alias{quickdraw}
\title{quickdraw dataset}
\format{
An object of class \code{data.frame} with 5000 rows and 785 columns.
}
\usage{
quickdraw
}
\description{
Imported from \code{datasets/quickdraw.csv}.
}
\keyword{datasets}
