% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{slasher}
\alias{slasher}
\title{slasher dataset}
\format{
An object of class \code{data.frame} with 485 rows and 2 columns.
}
\usage{
slasher
}
\description{
Imported from \code{datasets/slasher.csv}.
}
\keyword{datasets}
