% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{snack}
\alias{snack}
\title{snack dataset}
\format{
An object of class \code{data.frame} with 1709 rows and 12 columns.
}
\usage{
snack
}
\description{
Imported from \code{datasets/snack.csv}.
}
\keyword{datasets}
