% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{stock_market_2025}
\alias{stock_market_2025}
\title{stock_market_2025 dataset}
\format{
An object of class \code{data.frame} with 3650 rows and 14 columns.
}
\usage{
stock_market_2025
}
\description{
Imported from \code{datasets/stock_market_2025.csv}.
}
\keyword{datasets}
