% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{students}
\alias{students}
\title{students dataset}
\format{
An object of class \code{data.frame} with 100 rows and 7 columns.
}
\usage{
students
}
\description{
Imported from \code{datasets/students.csv}.
}
\keyword{datasets}
