% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{titanic dataset}
\format{
An object of class \code{data.frame} with 1000 rows and 8 columns.
}
\usage{
titanic
}
\description{
Imported from \code{datasets/titanic.csv}.
}
\keyword{datasets}
