% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{titanic_test}
\alias{titanic_test}
\title{titanic_test dataset}
\format{
An object of class \code{data.frame} with 308 rows and 8 columns.
}
\usage{
titanic_test
}
\description{
Imported from \code{datasets/titanic_test.csv}.
}
\keyword{datasets}
