% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-datasets.R
\docType{data}
\name{trash}
\alias{trash}
\title{trash dataset}
\format{
An object of class \code{data.frame} with 2000 rows and 12 columns.
}
\usage{
trash
}
\description{
Imported from \code{datasets/trash.csv}.
}
\keyword{datasets}
